Modification of the architecture of user routes
