+++ pages/users/new-user.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+type Props = {};
+
+function NewUser({}: Props) {
+  return <div>NewUser</div>;
+}
+
+export default NewUser;
