--- src/interfaces/interfaces.ts
@@ -1,11 +1,0 @@
-export interface IUser {
-  id: string;
-  firstname: string;
-  lastname: string;
-  username: string;
-  email: string;
-  imageUrl: string | null;
-  role: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
