+++ src/types/types.ts
@@ -7,3 +7,15 @@ export type TnewUser = {
   role: string;
   password: string;
 };
+
+export type TUser = {
+  id: string;
+  firstname: string;
+  lastname: string;
+  username: string;
+  email: string;
+  imageUrl: string | null;
+  role: string;
+  createdAt: Date;
+  updatedAt: Date;
+};
